backend/requirements.txt
# API server
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
pydantic==2.6.3
python-dotenv==1.0.1

# Database
neo4j==5.17.0

# AI / LLM
google-genai==1.10.0

# Task generation
duckduckgo-search==6.2.1

# Resume parsing
pdfplumber==0.11.0

# Web scraping
playwright==1.44.0
requests==2.31.0

# Fuzzy matching
thefuzz==0.22.1
